ICV_HLINE is split into several specific cases, according to pixel_size,
to optimize memory copies of the same color components along the line.