@vpisarev 2017-02-21
@chacha21, it breaks Android build and overall looks overcomplicated. Can we come up with something more compact, which does not require all those little-endian/big-endian tricks?