@chacha21 2017-02-21
I don't have a local android toolchain to test, so I am bound to submit to OpenCV buildbot to see the results. However Android seems to not be scheduled by default.

The endianess burden is necessary, but in some way it reveals that OpenCV lacks a standard information about endianess. A boost-like endian.h would be useful.

About something more compact, I don't know how to improve it.