@alalek 2017-03-01
It is better to use something like this:
```
template <unsigned pix_size_forced /*= 0*/>
static inline void icv_hline_impl(uchar* ptr, size_t xl, size_t xr, const uchar* color, unsigned pix_size_)
{
    const unsigned pix_size = pix_size_forced ? pix_size_forced : pix_size_;

    uchar* hline_ptr = ptr + xl*pix_size;
    uchar* hline_max_ptr = ptr + xr*pix_size;

    for ( ; hline_ptr <= hline_max_ptr; hline_ptr += pix_size)
    {
        for (unsigned c = 0; c < pix_size; c++)
        {
            hline_ptr[c] = color[c];
        }
    }
}

#define ICV_HLINE( ptr, xl, xr, color, pix_size ) \
{ \
    if (pix_size == 1) \
        icv_hline_impl<1>((uchar*)ptr, (xl), (xr), (const uchar*)color, /*unused*/pix_size); \
    else if (pix_size == 3) \
        icv_hline_impl<3>((uchar*)ptr, (xl), (xr), (const uchar*)color, /*unused*/pix_size); \
    else if (pix_size == 4) \
        icv_hline_impl<4>((uchar*)ptr, (xl), (xr), (const uchar*)color, /*unused*/pix_size); \
    else \
        icv_hline_impl<0>((uchar*)ptr, (xl), (xr), (const uchar*)color, pix_size); \
}
```

During template instantiation with fixed "pix_size" compiler will be able to unroll/optimize inner loop.