@chacha21 2017-03-01
The timings show that my approach still largely outperforms a simple loop unrolling
With a reproducible loop of drawing commands executed several times :

Current OpenCV 3.2 implementation :
     1 channels : 6780 ms
     3 channels : 12586 ms
     4 channels :  17671 ms

Using the template to unroll inner loop
     1 channels : 2458 ms
     3 channels :  5765 ms
     4 channels :  7581 ms

With my complex packaging as 32 bit integers
     1 channels : 384 ms
     3 channels :  1972 ms
     4 channels :  1149 ms