@chacha21 2017-03-03
Added some cleanup
Here are the current timings I get :

(1) : OpenCV 3.2.2
(2) : using only ICV_HLINE_X
(3) : using other variants of ICV_HLINE for element sizes <= 8
(4) : using all variants of ICV_HLINE up to element_size 32

                                                        (1)          (2)          (3)          (4)
type( 0(depth),1(channels)) =  0, elem_size =  1;    6816.27ms     851.99ms     394.18ms     392.61ms
type( 0(depth),3(channels)) = 16, elem_size =  3;   12600.38ms    1751.10ms    1494.44ms    1631.90ms
type( 0(depth),4(channels)) = 24, elem_size =  4;   19770.58ms    2388.67ms    1168.42ms    1162.85ms
type( 1(depth),1(channels)) =  1, elem_size =  1;    6822.77ms     851.84ms     390.28ms     392.47ms
type( 1(depth),3(channels)) = 17, elem_size =  3;   12610.19ms    1770.99ms    1475.96ms    1640.29ms
type( 1(depth),4(channels)) = 25, elem_size =  4;   19794.15ms    2374.06ms    1146.02ms    1160.07ms
type( 2(depth),1(channels)) =  2, elem_size =  2;    9967.02ms    1304.84ms     694.41ms     709.82ms
type( 2(depth),3(channels)) = 18, elem_size =  6;   27573.70ms    3531.09ms    2878.40ms    3047.34ms
type( 2(depth),4(channels)) = 26, elem_size =  8;   35305.18ms    4326.99ms    1968.01ms    1969.80ms
type( 3(depth),1(channels)) =  3, elem_size =  2;    9969.13ms    1307.69ms     709.25ms     709.69ms
type( 3(depth),3(channels)) = 19, elem_size =  6;   27585.40ms    3538.09ms    2885.79ms    3028.01ms
type( 3(depth),4(channels)) = 27, elem_size =  8;   35313.23ms    4310.90ms    1955.32ms    1966.05ms
type( 4(depth),1(channels)) =  4, elem_size =  4;   19776.41ms    2380.82ms    1150.15ms    1155.54ms
type( 4(depth),3(channels)) = 20, elem_size = 12;   50826.36ms    6338.50ms    6395.28ms    6575.13ms
type( 4(depth),4(channels)) = 28, elem_size = 16;   66306.04ms    8162.33ms    8189.29ms    7072.16ms
type( 5(depth),1(channels)) =  5, elem_size =  4;   19787.27ms    2367.58ms    1153.50ms    1203.38ms
type( 5(depth),3(channels)) = 21, elem_size = 12;   50847.69ms    6343.89ms    6371.89ms    6568.66ms
type( 5(depth),4(channels)) = 29, elem_size = 16;   66309.04ms    8156.95ms    8183.65ms    7086.30ms
type( 6(depth),1(channels)) =  6, elem_size =  8;   35332.65ms    4306.21ms    1967.08ms    1967.89ms
type( 6(depth),3(channels)) = 22, elem_size = 24;   97278.33ms   12051.37ms   12114.59ms   10695.41ms
type( 6(depth),4(channels)) = 30, elem_size = 32;  128364.84ms   15763.29ms   15774.99ms   14254.50ms