@vpisarev 2017-05-04
@chacha21, sorry for a huge delay. The current version of the patch is enormously big, even though it's not a new vision algorithm or other critical function, it's visualization function. We found that OpenCV grew very big, people complain about the size from time to time, so let's optimize not only for the speed, but for the code size as well. Besides, it's not clear whether the patch gives any acceleration not for rectangles, but, say, for think lines, ellipses or unfilled rectangles with thick outlines? e.g.

<pre>
for( int i = 0; i < 100000; i++ )
{
      Point pt1(rand()%img.width, rand()%img.height);
      Point pt2(rand()%img.width, rand()%img.height);
      line(img, pt1, pt2, Scalar(rand()%256, rand()%256, rand()%256), (rand()%4)*2+1);
}
</pre>

I think, we need to unroll the patch back to some earlier modifications. It's clear that we need to use memset() in some cases, and this is the main point of your patch, but it's also clear that for this particular task the patch should be much more compact and should not degrade performance in the case of other drawn primitives, like think lines.