@vpisarev 2017-05-23
@chacha21, the performance with the latest version is decent! thank you a lot for submitting and refining the patch :+1: